+++ packages/components/src/ProForm/hooks/useForm.tsx
@@ -1,0 +1,15 @@
+import { useRef } from 'react';
+
+const useForm = () => {
+  const clickRef = useRef<any>();
+
+  // 验证表单
+  const submitvalidate = () => clickRef.current.click();
+
+  return {
+    clickRef,
+    submitvalidate,
+  };
+};
+
+export default useForm;